my_django18_project/requirements.txt
django-markdown-deux==1.0.5
markdown2==2.3.9
